Feat: Add data handling utils
