+++ src/routes/Dashboard/ScoreListChart/types/index.ts
@@ -1,0 +1,6 @@
+export interface IScoreList {
+  year: string
+  score: number
+  fill: string
+  dotFill: string
+}
